requirements.txt
streamlit
pandas
supabase
scikit-learn
joblib
pytoml
matplotlib